Move the data transfer objects and slideshow template to their own package; in preparation for package extraction
