+++ webapp/core/src/main/java/com/hp/autonomy/frontend/reports/powerpoint/dto/ListData.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
  */
 
-package com.hp.autonomy.frontend.find.core.export;
+package com.hp.autonomy.frontend.reports.powerpoint.dto;
 
 import lombok.Data;
